requirements.txt
atproto==0.0.39
vaderSentiment==3.3.2
python-dotenv==1.0.1 